requirements.txt
affine==2.4.0
asgiref==3.8.1
attrs==23.2.0
certifi==2024.7.4
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.2.1
cycler==0.12.1
Django==5.0.7
fonttools==4.53.1
image==1.5.33
kiwisolver==1.4.5
matplotlib==3.9.1
numpy==2.0.1
packaging==24.1
pillow==10.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
rasterio==1.3.10
setuptools==71.1.0
six==1.16.0
snuggs==1.4.7
sqlparse==0.5.1
